Move logger into config folder
